byllmEMMA/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
byllm==0.4.6
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fastuuid==0.14.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
jaclang==0.9.0
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
litellm==1.80.0
loguru==0.7.3
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numpy==2.2.6
ollama==0.6.1
openai==2.8.0
opencv-python==4.12.0.88
packaging==25.0
pillow==10.4.0
propcache==0.4.1
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.29.0
safetensors==0.6.2
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
zipp==3.23.0
